requirements.txt
pinecone-client==2.2.4
pandas==2.1.3
python-dotenv==1.0.0 